+++ app/src/main/java/com/poupa/vinylmusicplayer/helper/MusicPlayerRemote.java
@@ -367,7 +367,7 @@ public class MusicPlayerRemote {
         }
     }
 
-    public static void setShuffleMode(final int shuffleMode) {
+    private static void setShuffleMode(final int shuffleMode) {
         if (musicService != null) {
             musicService.setShuffleMode(shuffleMode);
         }