+++ app/src/main/java/com/poupa/vinylmusicplayer/model/smartplaylist/AbsSmartPlaylist.java
@@ -36,7 +36,7 @@ public abstract class AbsSmartPlaylist extends AbsCustomPlaylist {
     public boolean canImport() {return false;}
 
     public void importPlaylist(@NonNull Context context, @NonNull Playlist playlist) {
-        // Notify app of clear event, so that the smart playlists are refreshed
+        // Notify app of the event, so that the smart playlists are refreshed
         if (canImport()) context.sendBroadcast(new Intent(MusicService.META_CHANGED));
     }