+++ app/src/main/java/com/poupa/vinylmusicplayer/ui/fragments/player/flat/FlatPlayerFragment.java
@@ -151,13 +151,6 @@ public class FlatPlayerFragment extends AbsPlayerFragment implements SlidingUpPa
         updateQueue();
     }
 
-    @Override
-    public void onMediaStoreChanged() {
-        // TODO If a song is removed from the MediaStore, this is not reflected in the playing queue untill restart
-        // TODO If a song is updated (tags edited), this is not reflected in the playing queue until restart
-        updateQueue();
-    }
-
     private void updateQueue() {
         playingQueueAdapter.swapDataSet(MusicPlayerRemote.getPlayingQueue(), MusicPlayerRemote.getPosition());
         playerQueueSubHeader.setText(MusicPlayerRemote.getQueueInfoString());
@@ -375,7 +368,7 @@ public class FlatPlayerFragment extends AbsPlayerFragment implements SlidingUpPa
         @Override
         public void updateCurrentSong(@NonNull final Song song) {
             currentSong = song;
-            currentSongViewHolder.title.setText(song.title);
+            currentSongViewHolder.title.setText(song.getTitle());
             currentSongViewHolder.text.setText(MusicUtil.getSongInfoString(song));
 
             if (PreferenceUtil.getInstance().animatePlayingSongIcon()) {
@@ -411,7 +404,7 @@ public class FlatPlayerFragment extends AbsPlayerFragment implements SlidingUpPa
 
         @Override
         public void updateCurrentSong(@NonNull final Song song) {
-            fragment.toolbar.setTitle(song.title);
+            fragment.toolbar.setTitle(song.getTitle());
             fragment.toolbar.setSubtitle(MusicUtil.getSongInfoString(song));
         }